A lot of changes. First:
All the data labeling and information is synced to be as in the proposed
visualization scheme.
A  code list was used to standardize labels in all the pertinent files, for
NAICS, SOC and CIP. These are part of the Master Scripts.

For indicators with tables, the total was added as an additional value.

In this commit indicators have been tested up to Indicator 011. The next commit
will push the latest tested datasets for the remaining indicators.
All changed datasets have a "y" as a suffix before
the year. The original data is available from the files without the "y".

CIP codes were added to indicators 12,25 and 26.
Indicator 27 was relabeled.
Some vectors were converted to numeric in order to perform totals, which may
have changed their formatting (Indicator 011).

Bugs were fixed in Indicator 015, and the Indicator 014 label for "% without
High Speed Internet" was changed to "% with Low-Speed Internet"