A lot of changes:
Files were ran with an automated script to find further bugs. Whenever possible, these were fixed.
Indicator012 .xlsx had an error in which the information for the Master's was the same for the Doctor. The R code did run properly,i.e., the results were as expected. Please change.

Indicator010 was reworked to be melted into long-data form. Its components were split by concept, so no manual intervention is required.

Some total rows were added to indicators, where deemed appropriate.

The files which had multi-sheet datasets had to be broken into separate files, in order for the
automated script to run.

Whenever possible, descriptions were standardized using the Manuals for the SOC 2010, NAICS 2012 and CIP 2010 code files.

Some renaming of rows and columns occurred, in order to comply with the charts/tables in the Excel File.

Indicators 002 and 023 ran into a bug, due to a corrupted database. The original totals are the correct ones, however: 2,523 firms for 2015-Q1, 68,331 employees in 2015-Q1-m1, etc. This is corrected in the latter commit.